# api/description/api.yml
openapi: 3.0.0
info:
  title: Lister
  description: Save your list
  version: "0.1"
servers: 
  - url: "/" 
paths: 
  /auth/register:
    post: 
      summary: Register a new account
      operationId: authRegister
      requestBody: 
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
      responses: 
        '200':
          description: OK
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/AuthRegisterResponse"
        '422':
          description: validation error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/ValidationError"
        '4xx':
          description: error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

  /auth/login:
    post: 
      summary: Login
      description: Returns JWT token on success
      operationId: authLogin
      requestBody: 
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/AuthLoginRequest"
      responses: 
        '200':
          description: OK
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/AuthLoginResponse"
        '422':
          description: validation error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/ValidationError"
        '4xx':
          description: error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

  /user:
    get:
      summary: Get authenticated user
      responses: 
        '200':
          description: user
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/User"
        '4xx':
          description: error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
            
  /user/lists:
    post: 
      summary: Create new list
      operationId: userCreateList
      requestBody: 
        content: 
          application/json:
            schema: 
              $ref: "#/components/schemas/ListCreateRequest"
      responses: 
        '201':
          description: Created
        '4xx':
          description: error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
    get: 
      summary: Get user's lists
      operationId: userLists
      responses: 
        '200':
          description: OK
          content: 
            application/json:
              schema: 
                type: array
                items: 
                  $ref: "#/components/schemas/List"
        '4xx':
          description: error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
  /user/lists/{list_id}:
    patch: 
      summary: Update list
      operationId: userListUpdate
      parameters: 
        - $ref: "#/components/parameters/pathListID"
      requestBody: 
        content: 
          application/json:
            schema: 
              $ref: "#/components/schemas/ListUpdateRequest"
      responses: 
        '200':
          description: OK
        '4xx':
          description: error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
    delete: 
      summary: Delete a list
      operationId: userDeleteList
      parameters: 
        - $ref: "#/components/parameters/pathListID"
      responses: 
        '200':
          description: OK
        '4xx':
          description: error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
  /user/lists/{list_id}/items:
    get:
      summary: Get items of a List
      operationId: listItems
      parameters:
        - $ref: "#/components/parameters/pathListID"
      responses: 
        '200':
          description: OK
          content: 
            application/json:
              schema: 
                type: array
                items: 
                  $ref: "#/components/schemas/ListItem"
    post: 
      summary: Add item to list
      operationId: listAddItem
      parameters: 
        - $ref: "#/components/parameters/pathListID"
      requestBody: 
        content: 
          application/json:
            schema: 
              $ref: "#/components/schemas/AddItemToListRequest"
      responses: 
        '201':
          description: Created
        '4xx':
          description: error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
  /user/lists/{list_id}/items/{item_id}:
    patch:
      summary: Update one item of a List
      operationId: updateListItem
      parameters:
        - $ref: "#/components/parameters/pathListID"
        - $ref: "#/components/parameters/pathItemID"
      requestBody:
        content: 
          application/json:
            schema: 
              $ref: "#/components/schemas/ListItemUpdateRequest"
      responses: 
        '200':
          description: OK
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/ListItem"                
        '4xx':
          description: error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"        
    delete: 
      summary: Delete an item
      operationId: deleteListItem
      parameters: 
        - $ref: "#/components/parameters/pathListID"
        - $ref: "#/components/parameters/pathItemID"
      responses: 
        '200':
          description: OK
        '4xx':
          description: error
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"        

security: []
components: 
  schemas: 
    Error:
      required: [code, message]
      properties: 
        code:
          type: string
        message:
          type: string
    ValidationError:
      required: [code, message, params]
      properties: 
        code:
          type: string
        message:
          type: string 
        params:
          type: array
          items: 
            $ref: "#/components/schemas/ValidationErrorParams"
    ValidationErrorParams:
      type: object
        
    AuthRegisterRequest:
      $ref: components/schemas/auth-register-request.yml
    AuthRegisterResponse:
      $ref: components/schemas/auth-register-response.yml
    AuthLoginRequest:
      $ref: components/schemas/auth-login-request.yml
    AuthLoginResponse:
      $ref: components/schemas/auth-login-response.yml
    User:
      type: object
      required: [id, name, email, created_at, updated_at]
      properties: 
        id:
          type: integer
          x-go-type: int64
        name:
          type: string
        email:
          type: string
        created_at:
          type: string
          x-go-type: time.Time
        updated_at:
          type: string
          x-go-type: time.Time
    List:
      type: object
      required: [id, title, description, created_at, updated_at]
      properties: 
        id:
          type: integer
          x-go-type: int64
        title:
          type: string
        description:
          type: string
        created_at:
          type: string
          x-go-type: time.Time
        updated_at:
          type: string
          x-go-type: time.Time
    ListItem:
      type: object
      required: [id, body, created_at, updated_at]
      properties: 
        id:
          type: integer
          x-go-type: int64
        body:
          type: string
        created_at:
          type: string
          x-go-type: time.Time
        updated_at:
          type: string
          x-go-type: time.Time
    ListCreateRequest:
      type: object
      required: [title, description]
      properties: 
        title:
          type: string
        description:
          type: string
    ListUpdateRequest:
      $ref: "#/components/schemas/ListCreateRequest"
    AddItemToListRequest:
      type: object
      required: [body] 
      properties: 
        body:
          type: string
    ListItemUpdateRequest:
      $ref: "#/components/schemas/AddItemToListRequest"
  parameters: 
    pathListID:
        in: path
        name: list_id
        schema: 
          type: integer
        required: true
        description: ID of the list
    pathItemID:
        in: path
        name: item_id
        schema: 
          type: integer
        required: true
        description: ID of the list item

  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT